+++ NEWS.rst
@@ -8,9 +8,13 @@ HappyBase 0.4 (not yet released)
 
 (TODO)
 
-* Documentation improvements
+* Add an optional `table_prefix_separator` argument to the
+  :py:class:`Connection` constructor, to specify the prefix used for the
+  `table_prefix` argument (`issue #3
+  <https://github.com/wbolster/happybase/issues/3>`_)
 * Add the Apache license conditions in the :doc:`license statement <license>`
   (for the included HBase parts)
+* Documentation improvements
 
 
 HappyBase 0.3